add: package.json base fix: workspace support for npm rename: shared -> libs

+++ app/web/src/App.tsx
@@ -1,11 +1,11 @@
 import { useState } from 'react'
+import { double } from '@foilctf/libs'
 import reactLogo from './assets/react.svg'
 import viteLogo from '/vite.svg'
 import './App.css'
 
 function App() {
   const [count, setCount] = useState(0)
-
   return (
     <>
       <div>
@@ -19,7 +19,7 @@ function App() {
       <h1>Vite + React</h1>
       <div className="card">
         <button onClick={() => setCount((count) => count + 1)}>
-          count is {count}
+          count is {double(count)}
         </button>
         <p>
           Edit <code>src/App.tsx</code> and save to test HMR
